Move TOTP modal trigger logic to FormCard component
Since this component is responsible for triggering the TOTP input modal on
the login page, the actual triggering logic belongs in it, rather than in
its parent page.
